# Speed up parseIPv4 3.3x.