```
name            old time/op    new time/op    delta
ParseIPv4-8       63.2ns ± 4%    20.1ns ± 4%  -68.17%  (p=0.000 n=19+19)
StdParseIPv4-8    93.9ns ± 3%    93.3ns ± 3%     ~     (p=0.211 n=20+19)
```

Signed-off-by: David Anderson <dave@natulte.net>